Map initialization, map state, point positions
Finnaly map is initialized at startup. Thanks to initialization algorithm knows where is the start point, end point, knows the number of colors and which colors are present on the game map. Added the mapState class which holds information about current map and every current point (or pipe), mapState was added to node - so now every node holds this object