Updated the Component Labeling
In come cases equivalent labels didn't work, but now after changes detecting stranded point or sector is working (in few tested cases), many functions are duplicated and tbh it's a little bit chaotic but i first want to make sure it is even working. Component Labeling is written without Disjoint-set data structure - if deadlines allows or it won't work in the end it will be rewritten.