Validating data, counting points, time fix, game start fix
Data sent by client to the server is now validated (atm. only checking start/end points placement and size of the map. Points are counted and displayed after map is changed to the next one (up to changes), Time is now corectly displayed on both clients. Two players are needed to start a match, otherwise error will be displayed