Fixed a bunch of things, handling disconnect
Fixed a bit forced moves, they can detect if it is not a forced move faster. Removed extra for loop in moves. Handled user disconnects from room - when game hasn't started yet, alert is shown and he is removed from room. When game is started and user disconnects, game is stopped and user who is still in game is a winner